Portal admin refactoring
