+++ packages/portal-admin/package.json
@@ -66,9 +66,7 @@
     "react-hook-form": "^7.48.2",
     "react-router-dom": "^6.19.0",
     "single-spa": "^5.9.3",
-    "single-spa-react": "^4.3.1",
-    "yup": "^1.3.3",
-    "zod": "^3.22.4"
+    "single-spa-react": "^4.3.1"
   },
   "types": "dist/jembi-portal-admin.d.ts"
 }
