+++ packages/portal-admin/src/components/FormFieldsComponents/FormInputRadio.tsx
@@ -7,7 +7,7 @@ import {
   RadioGroup
 } from '@mui/material'
 import { Controller } from 'react-hook-form'
-import { FormInputProps } from './FormInputProps'
+import { FormInputProps } from '../FormInputProps'
 import { useState } from 'react'
 
 export const FormInputRadioGroup: React.FC<FormInputProps> = ({
