+++ packages/portal-admin/src/declarations.d.ts
@@ -40,4 +40,3 @@ declare module '*.svg' {
 
 declare module '@jembi/openhim-core-api';
 declare module '@jembi/openhim-theme';
-declare module 'zod'
